refactor and move auth options to lib
