--- src/components/Providers.js
@@ -1,6 +1,0 @@
-"use client";
-import { SessionProvider } from "next-auth/react";
-
-export default function Providers({ children }) {
-  return <SessionProvider>{children}</SessionProvider>;
-}
