+++ src/components/SignOut.js
@@ -1,0 +1,3 @@
+export default function SignOut() {
+  return <a href="/api/auth/signout">Sign Out</a>;
+}
